+++ src/app/shared/models/dices.model.ts
@@ -22,7 +22,7 @@ export class DiceRoll {
     if (roll && roll.dice === 0) {return `${roll.modifier}`;}
     else if (!roll) {return '';}
     const modifier = roll.modifier ? `${roll.modifier}` : '';
-    const sign = roll.modifier > 0 ? '+' : roll.modifier < 0 ? '-' : '';
+    const sign = roll.modifier > 0 ? '+' : roll.modifier < 0 ? '' : '';
     return `${roll.dice}D${roll.sides}${sign}${modifier}`;
   }
 }